Persist launcher window geometry and improve display/resolution handling
### Motivation

- Persist main window geometry in the unified settings store so window position/size survive across platforms and no longer rely on `QSettings`/platform-specific storage.
- Ensure restored windows are visible on currently connected screens and save geometry when displays are removed.
- Prevent stale / oversized windowed resolutions and scaling ranges when switching to different displays by validating display index and desktop resolution.

### Description

- Added `launcher.mainWindow.geometry` to `config/schemas/settings.json` with `valid`, `x`, `y`, `width`, and `height` fields to store main window geometry.
- Reworked `MainWindow` to load geometry from `settings` and replaced `QSettings` usage, added `ensureWindowVisibleOnExistingScreen()` and `saveWindowSettings()`, and connected `screenRemoved` to trigger saving the geometry.
- Updated destructor to call `saveWindowSettings()` and removed dependency on `CLauncherDirs` for window settings.
- Enhanced `CSettingsView` to validate and clamp `video.displayIndex`, to use the selected display when computing preferred rendering resolution, added `findDesktopResolution()` and logic to include a custom windowed resolution only if it fits the selected display, and to update `settings[

+++ launcher/mainwindow_moc.cpp
@@ -197,10 +197,10 @@ void MainWindow::saveWindowSettings()
 	//save window settings
 	Settings windowGeometry = settings.write["launcher"]["mainWindow"]["geometry"];
 	windowGeometry["valid"].Bool() = true;
-	windowGeometry["x"].Float() = pos().x();
-	windowGeometry["y"].Float() = pos().y();
-	windowGeometry["width"].Float() = size().width();
-	windowGeometry["height"].Float() = size().height();
+	windowGeometry["x"].Integer() = pos().x();
+	windowGeometry["y"].Integer() = pos().y();
+	windowGeometry["width"].Integer() = size().width();
+	windowGeometry["height"].Integer() = size().height();
 #endif
 }
